Renamed rule `no-literals` -> `jsx-no-literals`
