+++ lib/rules/jsx-no-literals.js
